Add series titles, Add TitleGroup class
Add a #series method to return the series to which the object belongs.
The phrasing of a series is complicated, so :prepend and :append
properties are added to the #as_resource output of the TitleObject class

In order to make the representation of titles more consistent in
resources, change the output of CollectionObject#titles from a Hash
(essentially the "resource view") to a new TitleGroup class which is now
responsible for creating itself as a resource with :primary and :other
properties.