Add Movements to Collection Objects
Each object may be included in more than one movement, and #movements
will return an Array of Movement objects for each movement of which the
CollectionObject is a part.

The #as_resource method must add "no_objects" to the call to
Movements#as_resource to prevent the obvious infinite loop.