Qualify field in filter
Once we start interacting with the sites class through joins from the
CollectionObject class, 'siteid' becomes an ambiguous field in the SQL
behind the scenes because of the filter we've put on the dataset. It
could refer either to collection_tms_sites.siteid or
collection_objsitexrefs.siteid so it has to be qualified in the filter.