Add file to keep tests for odd cases
In the course of development we keep running into exceptional cases that
are rare but drive design decisions. For example, the one object that
belongs to two sites. The existence of this means object means that
CollectionObject#sites must return an array. We should keep track of
the cases that make these decisions necessary.