Add tests for assumptions
Sometimes design decisions are based on assumptions that may be true of
the currently available data but aren't certainly true for all data
present and future. For instance, there are no objects that belong to
more than one acquisition and there doesn't seem to be a reason to
assume that any object ever would belong to more than one acquisition
except that acquisitions can be funny things.

This file records and tests for those assumptions. If any of these tests
fail it would mean updating the API to reflect the unexpected change.