Add location field to objects
The location field is based on the exhobjxrefs table. That is, each
object is assigned a location for a given exhibition. To determine each
object's CURRENT location, the import process will look for exhibitions
that are currently running and update sort_fields.location with
exhobjxrefs.section for each object in those exhibitions.

A special "exhibitions" will have to be created for objects that are on
view but aren't in a particular temporary exhibition.

The exhobjxrefs.section field is a free text-entry field in TMS so some
way will have to be developed to check those values.