Make classes Collectible, Linkable
The CollectionObject class will need to provide methods that return
paginated resources of objects. The Collectible and Linkable modules
provide only instance methods since they have, before now, worked on
instances of other classes that were treated as containers of collection
objects -- so the paginated resource consisted of the objects belonging
to that instance.

CollectionObject class methods like #on_view will need to work without
an instance since they will return a selection of objects based on
criteria, not on another object so the methods in Collectible and
Linkable will have to work with instances OR classes.

The primary change is to create Linkable.make_links (class method) which
must be provided the clas name and other details as parameters.
Linkable#self_link (instance method) now just calls the class method
with the paramaters that can be determined from the instance.