Get defaults from Collectible in on_view
Because on_view is a class method and has to use the class method
version if Collectible#paginated_resource and also paginate its dataset
itself (when used as an instance method paginated_resource handles the
pagination) we need to get the default page and per_page values via
class methods as well. Get the values from Collectible so there is only
one method of handling the potential options and the default values.