Pass dataset as parameter to #paginated_resource
Make the dataset to be paginated a parameter of the #paginated_resource
method rather than an instance variable.

It turns out to not make much sense to keep track of the dataset to be
paginated as an instance variable that Collectible can act on,
especially once Collectible and Linkable methods need to be used from
class methods in CollectionObject.