Implement 'no_essay' option
Since the object essays are very long and not always needed, add a
'no_essay' option to #as_resource that, if set to 'true' (the string not
the boolean value), will suppress the :essay property in the resource.
This makes the :essay property OPTIONAL

Also add a REQUIRED :has_essay property that will be true if the object
has an essay, whether or not the 'no_essay' property has been set.