Use generic location in resource
We have given locations a mapping to a more generic version in the data.
Use this generic location in the resource. The resource has twp
properties, a site and a location. Remove id from resource because the
reported locatins are not unique, and there is currently no way to get a
location by id.