Allow JRuby failures
JRuby ain't so great with native extensions, so the sqlite3 gem won't
work with it. That's just for testing but I wouldn't bet on mysql
working so great either.

Now, are we going to use JRuby? No. Almost certainly never. But let's
leave it in the builds and just allow the failures. Maybe one day a
compatible gem will pop up!