Attempt to fix duplicate field
After update to new version that implemented the objects/recent
endpoint, that enpoint was returning

    Mysql::Error: Duplicate column name 'objectid'

This is probably the result of the same field being in the two joined
tables with "SELECT *". All tests passed so something is different in
the way Sqlite and Mysql deal with this.

Explicitly select only fields from collection_tms_objects in the hopes
that this resolves the error.